Split REST SPI package for better semver
