+++ modules/apps/oauth2-provider/oauth2-provider-rest-spi/src/main/resources/com/liferay/oauth2/provider/rest/spi/request/scope/checker/filter/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
